backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pymongo==4.3.3
python-dotenv==1.0.0
pydantic==1.10.7
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
websockets==11.0.1
scikit-learn==1.2.2
pandas==2.0.0
numpy==1.24.2

